Gensim_tokenization
So, this works for a string and a dictionary of strings (in Jupyter).  However,  I am looking into the API call.   I was also unsure if this should be one function or well,  how to break down the analysis into functions etc.   I think docstrings are going to be important in this one.